requirements.txt
baidu-aip==4.16.2
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
cycler==0.11.0
fonttools==4.31.2
idna==3.3
joblib==1.1.0
kiwisolver==1.4.2
matplotlib==3.5.1
numpy==1.21.5
packaging==21.3
pandas==1.3.5
Pillow==9.1.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
typing_extensions==4.1.1
urllib3==1.26.9
